Added a count down timer to the test Activity
Timer counts down the time left from the Current time to the ValidTo
time attribute of the Exam. Slightly changed the UI of the bottom bar.
All testing exams have a valid from date set to the current time and a
valid to date set to 10 min past the current time.